Restructured Code
Game State and Prjection Transformation now have their own wrapper classes.
This makes it so that you dont have to update a variable everywhere. Instead, store the wrapper everywhere and you can just change the variable contained inside the wrapper, and it is automatically updated everywhere. This is cuz how pointers work and whatever.